requirements.txt
numpy
pynrrd
itk